pyproject.toml
[project]
name = "CoL"
version = "0.1.0"
description = "description"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ichatbio-sdk~=0.2.1",
    "a2a-sdk[sqlite]",
    "python-dotenv~=0.9.9",
    "openai~=1.79.0",
    "instructor~=1.8.2",
    "pydantic~=2.11.4",
    "requests",
    "sqlalchemy",
    "aiosqlite",
]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]